pyproject.toml
[tool.poetry]
name = "ai-agent-experiments"
version = "0.1.0"
description = ""
authors = ["pavithrakamath"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.108.0"
requests = "^2.32.5"
python-dotenv = "^1.1.1"
mcp = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
